+++ Cargo.toml
@@ -101,7 +101,7 @@ markdown = "=1.0.0-alpha.21"
 mime_guess = "=2.0.5"
 nanoid = "=0.4.0"
 nest_struct = "=0.5.2"
-openidconnect = "=3.5.0"
+openidconnect = "=4.0.0"
 paginate = "=1.1.11"
 radarr-api-rs = "=3.0.1"
 rand = "=0.9.0-beta.0"
