+++ crates/utils/application/Cargo.toml
@@ -14,5 +14,10 @@ config = { path = "../../config" }
 file-storage-service = { path = "../../services/file-storage" }
 jwt-service = { path = "../../services/jwt" }
 media-models = { path = "../../models/media" }
+openidconnect = { workspace = true }
 reqwest = { workspace = true }
 sea-orm = { workspace = true }
+tracing = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["tracing"]
